requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1
redis==5.0.1
langchain==0.1.11
langchain-openai==0.1.1
langgraph==0.1.22
azure-ai-documentintelligence==1.0.0
azure-identity==1.15.0
azure-ai-vision==0.15.1b1
azure-storage-blob==12.19.0
pydantic==2.6.3
httpx==0.27.0
tenacity==8.2.3
